# packages/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: Stats App API
  version: "0.1.0"
  description: API versionada para análisis estadístico de fútbol

servers:
  - url: /api/v1
    description: API v1

paths:
  /:
    get:
      operationId: getRoot
      summary: Información del servicio
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceInfo"

  /health/live:
    get:
      operationId: getLiveness
      summary: Liveness probe
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /health/ready:
    get:
      operationId: getReadiness
      summary: Readiness probe
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessStatus"

  /matchdays/current:
    get:
      operationId: getCurrentMatchday
      summary: Jornada actual con partidos y cuotas Over/Under 2.5
      parameters:
        - name: matchday
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matchday"

  /matches/{match_id}:
    get:
      operationId: getMatchDetail
      summary: Detalle de un partido con estadísticas y cuotas
      parameters:
        - name: match_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetail"
        "404":
          description: Partido no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /matches/{match_id}/odds:
    get:
      operationId: getMatchOdds
      summary: Snapshots de cuotas de un partido
      parameters:
        - name: match_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Odds"
        "404":
          description: Partido no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /predictions/{prediction_id}:
    get:
      operationId: getPrediction
      summary: Recupera una predicción persistida por ID
      parameters:
        - name: prediction_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prediction"
        "404":
          description: Predicción no encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /opportunities:
    get:
      operationId: getOpportunities
      summary: Scanner de oportunidades (señales con edge, EV, calidad y riesgo)
      parameters:
        - name: min_edge
          in: query
          required: false
          schema:
            type: number
            minimum: -100
            maximum: 100
            default: 0
        - name: risk
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high]
        - name: matchday
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [edge, ev, probability]
            default: edge
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Opportunity"

  /parlays/estimate:
    post:
      operationId: estimateParlay
      summary: Estima cuota combinada, probabilidad y riesgo agregado de un parlay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParlayEstimateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParlayEstimate"
        "422":
          description: Selección no resoluble (sin predicción o sin cuota)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /model-versions:
    get:
      operationId: getModelVersions
      summary: Lista de versiones de modelo registradas
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModelVersion"

  /metrics:
    get:
      operationId: getMetrics
      summary: Métricas de evaluación (ROI, Brier, acierto y calibración)
      parameters:
        - name: model_version_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: threshold
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 30
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Metrics"

  /history:
    get:
      operationId: getHistory
      summary: Historial paginado de predicciones resueltas
      parameters:
        - name: model_version
          in: query
          required: false
          schema:
            type: string
        - name: result
          in: query
          required: false
          schema:
            type: string
            enum: [win, loss, void]
        - name: matchday
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryPage"

  /ops/metrics:
    get:
      operationId: getOpsMetrics
      summary: Métricas operativas en formato Prometheus (latencia, errores, cola y frescura)
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string

components:
  schemas:
    ServiceInfo:
      type: object
      required: [service, version, env]
      properties:
        service:
          type: string
          example: stats-api
        version:
          type: string
          example: "0.1.0"
        env:
          type: string
          example: development

    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok, degraded, unavailable]

    ReadinessStatus:
      type: object
      required: [status, checks]
      properties:
        status:
          type: string
          enum: [ok, degraded, unavailable]
        checks:
          type: object
          properties:
            postgres:
              type: string
            redis:
              type: string

    Team:
      type: object
      required: [id, name, short_name]
      properties:
        id:
          type: string
        name:
          type: string
        short_name:
          type: string

    Match:
      type: object
      required: [id, competition, kickoff_at, home_team, away_team, status]
      properties:
        id:
          type: string
        competition:
          type: string
        kickoff_at:
          type: string
          format: date-time
        home_team:
          $ref: "#/components/schemas/Team"
        away_team:
          $ref: "#/components/schemas/Team"
        status:
          type: string
          enum: [scheduled, finished]
        over_odds:
          type: number
          nullable: true
        under_odds:
          type: number
          nullable: true

    Matchday:
      type: object
      required: [matchday, updated_at, total_matches, matches]
      properties:
        matchday:
          type: integer
        updated_at:
          type: string
          format: date-time
        total_matches:
          type: integer
        matches:
          type: array
          items:
            $ref: "#/components/schemas/Match"

    TeamMatchStats:
      type: object
      required: [match_id, team_id, goals]
      properties:
        match_id:
          type: string
        team_id:
          type: string
        goals:
          type: integer
        shots:
          type: integer
          nullable: true
        shots_on_target:
          type: integer
          nullable: true
        possession:
          type: number
          nullable: true
        corners:
          type: integer
          nullable: true

    Odds:
      type: object
      required: [match_id, market, selection, decimal_odds, captured_at, provider]
      properties:
        match_id:
          type: string
        market:
          type: string
          enum: [over_under_2_5]
        selection:
          type: string
          enum: [over, under]
        decimal_odds:
          type: number
        captured_at:
          type: string
          format: date-time
        provider:
          type: string

    MatchDetail:
      type: object
      required: [match, stats, odds]
      properties:
        match:
          $ref: "#/components/schemas/Match"
        stats:
          type: array
          items:
            $ref: "#/components/schemas/TeamMatchStats"
        odds:
          type: array
          items:
            $ref: "#/components/schemas/Odds"
        predictions:
          type: array
          items:
            $ref: "#/components/schemas/Prediction"

    Prediction:
      type: object
      required: [id, match_id, model_version_id, market, selection, probability, fair_odds, data_quality, risk_level, inputs_hash, prediction_timestamp]
      properties:
        id:
          type: string
          format: uuid
        match_id:
          type: string
          format: uuid
        model_version_id:
          type: string
          format: uuid
        market:
          type: string
        selection:
          type: string
        probability:
          type: number
        fair_odds:
          type: number
        implied_probability:
          type: number
          nullable: true
        no_vig_probability:
          type: number
          nullable: true
        edge_pp:
          type: number
          nullable: true
        ev:
          type: number
          nullable: true
        data_quality:
          type: string
        risk_level:
          type: string
        inputs:
          type: string
          nullable: true
        inputs_hash:
          type: string
        prediction_timestamp:
          type: string
          format: date-time
        snapshot_age_minutes:
          type: integer
          nullable: true
        confidence_level:
          type: string
          nullable: true
        confidence_score:
          type: number
          nullable: true
        confidence_factors:
          type: array
          items:
            type: string
        explanation:
          type: object
          nullable: true
          properties:
            summary:
              type: string
            factors:
              type: array
              items:
                type: string
            risks:
              type: array
              items:
                type: string
            formula:
              type: string
            provenance:
              type: object
              properties:
                dataset:
                  type: string
                  nullable: true
                feature_set_version:
                  type: string
                  nullable: true
                inputs_hash:
                  type: string
            model_version:
              type: string
              nullable: true
            prediction_timestamp:
              type: string
            is_guarantee:
              type: boolean

    Opportunity:
      type: object
      required: [match_id, home_team_short, away_team_short, kickoff_at, market, selection, model_probability, market_no_vig_probability, observed_odds, fair_odds, edge_pp, ev, data_quality, risk_level, is_signal, signal_exclusions, snapshot_age_minutes]
      properties:
        match_id:
          type: string
        home_team_short:
          type: string
        away_team_short:
          type: string
        kickoff_at:
          type: string
          format: date-time
        market:
          type: string
        selection:
          type: string
          enum: [over, under]
        model_probability:
          type: number
        market_no_vig_probability:
          type: number
        observed_odds:
          type: number
        fair_odds:
          type: number
        edge_pp:
          type: number
        ev:
          type: number
        data_quality:
          type: string
        risk_level:
          type: string
        is_signal:
          type: boolean
        signal_exclusions:
          type: array
          items:
            type: string
        snapshot_age_minutes:
          type: integer
        confidence_level:
          type: string
        confidence_score:
          type: number
        confidence_factors:
          type: array
          items:
            type: string

    ErrorResponse:
      type: object
      required: [code, message, correlation_id]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        correlation_id:
          type: string

    ParlaySelectionRef:
      type: object
      required: [match_id, selection]
      properties:
        match_id:
          type: string
        market:
          type: string
          default: over_under_2_5
        selection:
          type: string
          enum: [over, under]

    ParlayEstimateRequest:
      type: object
      required: [selections]
      properties:
        selections:
          type: array
          minItems: 1
          maxItems: 3
          items:
            $ref: "#/components/schemas/ParlaySelectionRef"

    ResolvedSelection:
      type: object
      required: [key, match_id, market, selection, home_team_short, away_team_short, kickoff_at, probability, odds, fair_odds, edge_pp, data_quality, risk_level]
      properties:
        key:
          type: string
        match_id:
          type: string
        market:
          type: string
        selection:
          type: string
          enum: [over, under]
        home_team_short:
          type: string
        away_team_short:
          type: string
        kickoff_at:
          type: string
          format: date-time
        probability:
          type: number
        odds:
          type: number
        fair_odds:
          type: number
        edge_pp:
          type: number
        data_quality:
          type: string
        risk_level:
          type: string
        confidence_level:
          type: string
        confidence_score:
          type: number
        confidence_factors:
          type: array
          items:
            type: string

    ParlayEstimate:
      type: object
      required: [selections, combined_odds, naive_probability, estimated_probability, risk_level, risk_factors, correlation_warnings, assumes_independence, selection_count]
      properties:
        selections:
          type: array
          items:
            $ref: "#/components/schemas/ResolvedSelection"
        combined_odds:
          type: number
        naive_probability:
          type: number
        estimated_probability:
          type: number
        fair_combined_odds:
          type: number
          nullable: true
        risk_level:
          type: string
        risk_factors:
          type: array
          items:
            type: string
        correlation_warnings:
          type: array
          items:
            type: string
        assumes_independence:
          type: boolean
        selection_count:
          type: integer

    ModelVersion:
      type: object
      required: [id, name, version, status, feature_set_version, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
        status:
          type: string
        feature_set_version:
          type: string
        created_at:
          type: string
          format: date-time

    CalibrationBin:
      type: object
      required: [label, lower, upper, n]
      properties:
        label:
          type: string
        lower:
          type: number
        upper:
          type: number
        n:
          type: integer
        mean_predicted:
          type: number
          nullable: true
        observed_rate:
          type: number
          nullable: true

    Metrics:
      type: object
      required: [sample_size, wins, losses, voids, calibration_bins, sample_sufficient, threshold]
      properties:
        model_version_id:
          type: string
          nullable: true
        sample_size:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        voids:
          type: integer
        hit_rate:
          type: number
          nullable: true
        unit_roi:
          type: number
          nullable: true
        brier:
          type: number
          nullable: true
        calibration_bins:
          type: array
          items:
            $ref: "#/components/schemas/CalibrationBin"
        sample_sufficient:
          type: boolean
        threshold:
          type: integer

    HistoryItem:
      type: object
      required: [prediction_id, match_id, home_team_short, away_team_short, kickoff_at, market, selection, probability, model_version, prediction_timestamp, result, resolved_at]
      properties:
        prediction_id:
          type: string
          format: uuid
        match_id:
          type: string
        home_team_short:
          type: string
        away_team_short:
          type: string
        kickoff_at:
          type: string
          format: date-time
        market:
          type: string
        selection:
          type: string
        probability:
          type: number
        odds:
          type: number
          nullable: true
        model_version:
          type: string
        prediction_timestamp:
          type: string
          format: date-time
        result:
          type: string
          enum: [win, loss, void]
        resolved_at:
          type: string
          format: date-time

    HistoryPage:
      type: object
      required: [items, page, page_size, total, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/HistoryItem"
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer